// src/lib/generated/cerebro/ClientConfiguration.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ClientConfiguration = { runner_id: string, target_id: string, execution_module_id: string | null, binding_id: string | null, mcp_module_ids: Array<string>, mcp_enabled: boolean, module_labels: { [key in string]?: string }, grant_id: string | null, credential_expires_at: string | null, };