// src/lib/generated/cerebro/ConfigurationInput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MCPCapabilities } from "./MCPCapabilities"
import type { MCPScopeItem } from "./MCPScopeItem"

export type ConfigurationInput = {
  execution_module_id: string | null
  mcp_scope_modules: Array<MCPScopeItem>
  mcp_capabilities: MCPCapabilities
  mcp_enabled: boolean
}